+++ tests/epyccel/test_epyccel_lists.py
@@ -17,6 +17,18 @@ from pyccel import epyccel
 def language(request):
     return request.param
 
+@pytest.fixture( params=[
+        pytest.param("c", marks = pytest.mark.c),
+        pytest.param("fortran", marks = [
+            pytest.mark.skip(reason="lists not implemented in fortran"),
+            pytest.mark.fortran]),
+        pytest.param("python", marks = pytest.mark.python)
+    ],
+    scope = "module"
+)
+def stc_language(request):
+    return request.param
+
 def test_pop_last_element(language) :
     def pop_last_element():
         a = [1,3,45]
@@ -658,3 +670,23 @@ def test_extend_returned_list(language):
     epyc_f = epyccel(f, language=language)
     assert f() == epyc_f()
 
+def test_mutable_indexing(stc_language):
+    def f():
+        a = [1,2,3,4]
+        a[0] = 5
+        a[2] += 6
+        return a[0], a[1], a[2], a[3]
+
+    epyc_f = epyccel(f, language=stc_language)
+    assert f() == epyc_f()
+
+def test_mutable_multi_level_indexing(stc_language):
+    def f():
+        a = [1,2,3,4]
+        b = [a]
+        b[0][0] = 5
+        b[0][2] = 6
+        return a[0], a[1], a[2], a[3]
+
+    epyc_f = epyccel(f, language=stc_language)
+    assert f() == epyc_f()